move conversion functions between glm/phyx to separate file
